+++ CHANGELOG.md
@@ -425,3 +425,4 @@
 - 2024-06-07: Err integration, mod, summaries, site config, +6
 - 2024-06-08: Async mutate, mailbox, qualify, mod, +3
 - 2024-06-09: Health user, send, mutate,  outreach base, +4
+- 2024-06-10: Init leads, mailbox, bettercontact, descend, +2